Extracted sql to separated class
