+++ mod-source-record-manager-server/src/main/java/org/folio/dao/JobExecutionDaoImpl.java
@@ -30,29 +30,34 @@ import java.util.UUID;
 
 import static java.lang.String.format;
 import static java.util.Objects.nonNull;
-import static org.folio.dao.util.JobExecutionsColumns.COMPLETED_DATE_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.CURRENTLY_PROCESSED_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.ERROR_STATUS_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.FILE_NAME_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.HRID_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.ID_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.JOB_PROFILE_DATA_TYPE_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.JOB_PROFILE_ID_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.JOB_PROFILE_NAME_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.JOB_USER_FIRST_NAME_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.JOB_USER_LAST_NAME_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.PARENT_ID_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.PROFILE_SNAPSHOT_WRAPPER_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.PROGRESS_CURRENT_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.PROGRESS_TOTAL_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.SOURCE_PATH_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.STARTED_DATE_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.STATUS_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.SUBORDINATION_TYPE_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.TOTAL_COUNT_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.TOTAL_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.UI_STATUS_FIELD;
-import static org.folio.dao.util.JobExecutionsColumns.USER_ID_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.COMPLETED_DATE_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.CURRENTLY_PROCESSED_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.ERROR_STATUS_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.FILE_NAME_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.GET_BY_ID_SQL;
+import static org.folio.dao.util.JobExecutionDBConstants.GET_CHILDREN_JOBS_BY_PARENT_ID_SQL;
+import static org.folio.dao.util.JobExecutionDBConstants.GET_JOBS_NOT_PARENT_SQL;
+import static org.folio.dao.util.JobExecutionDBConstants.HRID_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.ID_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.INSERT_SQL;
+import static org.folio.dao.util.JobExecutionDBConstants.JOB_PROFILE_DATA_TYPE_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.JOB_PROFILE_ID_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.JOB_PROFILE_NAME_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.JOB_USER_FIRST_NAME_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.JOB_USER_LAST_NAME_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.PARENT_ID_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.PROFILE_SNAPSHOT_WRAPPER_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.PROGRESS_CURRENT_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.PROGRESS_TOTAL_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.SOURCE_PATH_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.STARTED_DATE_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.STATUS_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.SUBORDINATION_TYPE_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.TOTAL_COUNT_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.TOTAL_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.UI_STATUS_FIELD;
+import static org.folio.dao.util.JobExecutionDBConstants.UPDATE_SQL;
+import static org.folio.dao.util.JobExecutionDBConstants.USER_ID_FIELD;
 import static org.folio.rest.persist.PostgresClient.convertToPsqlStandard;
 
 /**
@@ -70,45 +75,6 @@ public class JobExecutionDaoImpl implements JobExecutionDao {
   private static final String TABLE_NAME = "job_execution";
   private static final String PROGRESS_TABLE_NAME = "job_execution_progress";
   public static final String GET_JOB_EXECUTION_HR_ID = "SELECT nextval('%s.job_execution_hr_id_sequence')";
-  private static final String GET_BY_ID_SQL = "SELECT * FROM %s WHERE id = $1";
-
-  private static final String INSERT_SQL =
-    "INSERT INTO %s.%s (id, hrid, parent_job_id, subordination_type, source_path, file_name, " +
-    "progress_current, progress_total, started_date, completed_date, status, ui_status, error_status, job_user_first_name, " +
-    "job_user_last_name, user_id, job_profile_id, job_profile_name, job_profile_data_type, job_profile_snapshot_wrapper) " +
-    "VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19, $20)";
-
-  public static final String UPDATE_SQL =
-    "UPDATE %s " +
-    "SET id = $1, hrid = $2, parent_job_id = $3, subordination_type = $4, source_path = $5, file_name = $6, " +
-    "progress_current = $7, progress_total = $8, started_date = $9, completed_date = $10, " +
-    "status = $11, ui_status = $12, error_status = $13, job_user_first_name = $14, job_user_last_name = $15, " +
-    "user_id = $16, job_profile_id = $17, job_profile_name = $18, job_profile_data_type = $19, " +
-    "job_profile_snapshot_wrapper = $20 " +
-    "WHERE id = $1";
-
-  private static final String GET_CHILDREN_JOBS_BY_PARENT_ID_SQL =
-    "WITH cte AS (SELECT count(*) AS total_count FROM %s " +
-    "WHERE parent_job_id = $1 AND subordination_type = 'CHILD') " +
-    "SELECT j.*, cte.*, p.jsonb -> 'total' total, " +
-    "(p.jsonb -> 'currentlySucceeded')::int + (p.jsonb -> 'currentlyFailed')::int currently_processed " +
-    "FROM %s j " +
-    "LEFT JOIN %s p ON  j.id = p.jobexecutionid " +
-    "LEFT JOIN cte ON true " +
-    "WHERE parent_job_id = $1 AND subordination_type = 'CHILD' " +
-    "LIMIT $2 OFFSET $3";
-
-  private static final String GET_JOBS_NOT_PARENT_SQL =
-    "WITH cte AS (SELECT count(*) AS total_count FROM %s " +
-    "WHERE subordination_type <> 'PARENT_MULTIPLE' AND %s) " +
-    "SELECT j.*, cte.*, p.jsonb -> 'total' total, " +
-    "(p.jsonb -> 'currentlySucceeded')::int + (p.jsonb -> 'currentlyFailed')::int currently_processed " +
-    "FROM %s j " +
-    "LEFT JOIN %s p ON  j.id = p.jobexecutionid " +
-    "LEFT JOIN cte ON true " +
-    "WHERE subordination_type <> 'PARENT_MULTIPLE' AND %s " +
-    "ORDER BY %s %s " +
-    "LIMIT $1 OFFSET $2";
 
   @Autowired
   private PostgresClientFactory pgClientFactory;
