--- mod-source-record-manager-server/src/main/java/org/folio/dao/util/JobExecutionsColumns.java
@@ -1,31 +1,0 @@
-package org.folio.dao.util;
-
-public final class JobExecutionsColumns {
-
-  public static final String ID_FIELD = "id";
-  public static final String HRID_FIELD = "hrid";
-  public static final String PARENT_ID_FIELD = "parent_job_id";
-  public static final String SUBORDINATION_TYPE_FIELD = "subordination_type";
-  public static final String JOB_PROFILE_ID_FIELD = "job_profile_id";
-  public static final String JOB_PROFILE_NAME_FIELD = "job_profile_name";
-  public static final String JOB_PROFILE_DATA_TYPE_FIELD = "job_profile_data_type";
-  public static final String PROFILE_SNAPSHOT_WRAPPER_FIELD = "job_profile_snapshot_wrapper";
-  public static final String SOURCE_PATH_FIELD = "source_path";
-  public static final String FILE_NAME_FIELD = "file_name";
-  public static final String PROGRESS_CURRENT_FIELD = "progress_current";
-  public static final String PROGRESS_TOTAL_FIELD = "progress_total";
-  public static final String STARTED_DATE_FIELD = "started_date";
-  public static final String COMPLETED_DATE_FIELD = "completed_date";
-  public static final String STATUS_FIELD = "status";
-  public static final String UI_STATUS_FIELD = "ui_status";
-  public static final String ERROR_STATUS_FIELD = "error_status";
-  public static final String JOB_USER_FIRST_NAME_FIELD = "job_user_first_name";
-  public static final String JOB_USER_LAST_NAME_FIELD = "job_user_last_name";
-  public static final String USER_ID_FIELD = "user_id";
-  public static final String TOTAL_COUNT_FIELD = "total_count";
-  public static final String CURRENTLY_PROCESSED_FIELD = "currently_processed";
-  public static final String TOTAL_FIELD = "total";
-
-  private JobExecutionsColumns() {
-  }
-}
